requirements.txt
flask==2.0.1
werkzeug==2.0.3
numpy==1.22.4
pandas==1.4.3
openai==0.28.1
python-dotenv==1.0.0
python3-discogs-client==2.6.0
requests==2.28.2
gunicorn==20.1.0
Jinja2==3.0.3
MarkupSafe==2.1.1
markdown==3.4.3 
